enum typo split test, remove unused setup methods in tests
